tidied up main.js using ES6
