+++ .gitignore
@@ -13,4 +13,4 @@ languages/__pycache__/
 ._.DS_Store
 **/.DS_Store
 **/._.DS_Store
-
+qodana.yaml
