+++ package.json
@@ -3,6 +3,7 @@
   "version": "1.1.0",
   "description": "A language-learning app that knows which words you know for faster learning by immersion",
   "main": "main.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "electron ."
